Added description to configs

+++ lib/fastlane/plugin/twine/actions/twine_check_action.rb
@@ -5,20 +5,25 @@ module Fastlane
         require 'diffy'
         Actions.verify_gem!('twine')
         Helper::TwineConfigParser.parse(params).each do |config|
-          Actions.sh(Helper::TwineCommandHelper.generate_command(
-                       params[:use_bundle_exec] && shell_out_should_use_bundle_exec?,
-                       config.source_path,
-                       config.destination_path + '.tmp',
-                       config.twine_args
-          ), error_callback: lambda { |e|
-                               UI.user_error!('Error while generating localization file ' + destination_path)
-                             })
-          files_are_the_same = FileUtils.compare_file(config.destination_path, config.destination_path + '.tmp')
-          if params[:visualize_diff] && !files_are_the_same
-            puts Diffy::Diff.new(config.destination_path, config.destination_path + '.tmp', source: 'files').to_s(:color)
+          if !File.exist?(config.source_path)
+            UI.user_error!("Source file " + config.source_path + " not found")
+          else
+            UI.message(config.description)
+            Actions.sh(Helper::TwineCommandHelper.generate_command(
+                         params[:use_bundle_exec] && shell_out_should_use_bundle_exec?,
+                         config.source_path,
+                         config.destination_path + '.tmp',
+                         config.twine_args
+            ), error_callback: lambda { |e|
+                                 UI.user_error!('Error while generating localization file ' + config.destination_path)
+                               })
+            files_are_the_same = FileUtils.compare_file(config.destination_path, config.destination_path + '.tmp')
+            if params[:visualize_diff] && !files_are_the_same
+              puts Diffy::Diff.new(config.destination_path, config.destination_path + '.tmp', source: 'files').to_s(:color)
+            end
+            File.delete(config.destination_path + '.tmp')
+            UI.test_failure!("Files are different") unless files_are_the_same
           end
-          File.delete(config.destination_path + '.tmp')
-          UI.test_failure!("Files are different") unless files_are_the_same
         end
       end
